Commenting and Cleanup
Commented all functions and build new helper file `tile.dart` which can be imported into files that use any Arc or Task tiles.
